finalizando

+++ src/main/java/com/campana_backend/campana_backend/repository/ClienteRepository.java
@@ -2,15 +2,20 @@
 
 package com.campana_backend.campana_backend.repository;
 
+
+
 import com.campana_backend.campana_backend.models.Cliente;
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long>{
+public interface ClienteRepository extends JpaRepository<Cliente, Integer>{
     
     
     List<Cliente> findByCedula(String cedula);
+    
+    
 
 }
